+++ attack.js
@@ -1,0 +1,3 @@
+export default function attack() {
+    console.log(`${this.name} fight`);
+}
