+++ createElement.js
@@ -1,0 +1,7 @@
+const createElement = (tag, className) => {
+    const $tag = document.createElement(tag);
+    if (className) { $tag.classList.add(className); }
+    
+    return $tag;
+}
+export default createElement;
