+++ elHp.js
@@ -1,0 +1,4 @@
+export default function elHp() {
+    const $playerLife = document.querySelector('.player'+ this.player + ' .life');
+    return $playerLife;
+}
