+++ randomDamage.js
@@ -1,0 +1,2 @@
+const randomDamage = (damage) => Math.ceil(Math.random() * damage);
+export default randomDamage;
