+++ renderHp.js
@@ -1,0 +1,5 @@
+import elHp from "./elHp.js";
+export default function renderHp() {
+    const $playerLife = this.elHp();
+    $playerLife.style.width = `${this.hp}%`;
+}
