+++ utils.js
@@ -1,0 +1,4 @@
+export const $arenas = document.querySelector('.arenas');
+export const $randomButton = document.querySelector('.button');
+export const $formFight = document.querySelector('.control');
+export const $chat = document.querySelector('.chat');
